# Reduce `Client.format` allocations