--- a/statsd/statsd_benchmark_test.go
+++ b/statsd/statsd_benchmark_test.go
@@ -1,3 +1,5 @@
+// +build go1.7
+
 package statsd
 
 import (