@vcabbage 2018-03-04
CI failed because I used [`B.Run()`](https://golang.org/pkg/testing/#B.Run), which was added in Go 1.7.

Since they're benchmarks and not used by consumers of the library or during CI, I added a build tag to omit them when compiled with < Go 1.7. I could also not use that API, but it does make the benchmarks nicer.