@zippolyte 2018-04-26
Hi @vcabbage, thanks for your interest !

While it is true this improves performance a bit, I don't feel like the trade off is worth it. 
You end up with a less flexible API, where you'd also need to add some checking mechanism to avoid panics caused by a buffer too small. Additionally, this makes the code a little less readable, not great for maintainability and contributing.
If there was no trade off to this improvement, I would definitely take it. Right now though, I am not really in favor of merging it.