@gzussa 2018-07-02
Hi @vcabbage, This is great. Thanks for making this improvement :).
@ofek Looks good to me